long -> int32 for linkers/*/x.out.h
This solves the segfault when linking principia/compilers/hoc
which happened probably when 8l was linking libc.a(sin.c)
which contained float and because of the use of sprint()
in 8l/obj.c sprint() probably overflowed and corrupted the memory.

Note that we should get rid of sprint too!

+++ src/lib9/fmt/sprint.c
@@ -18,7 +18,8 @@ sprint(char *buf, char *fmt, ...)
 	 * optimizes the test away.  casting to uintptr works around this bug.
 	 */
 	if((uintptr)buf+len < (uintptr)buf)
-		len = -(uintptr)buf-1;
+		len = (uint)-(uintptr)buf-1;
+    //print("sprint: len = %d\n", len);
 
 	va_start(args, fmt);
 	n = vsnprint(buf, len, fmt, args);
